+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Allow configuring various shadow/fast reflection options for vactube object ents.
 * Boolean keyvalues (yes/no) can now be set to `!$var`, to invert the value of the variable. 
 * Propcombine will now preserve prop fade distances, by calculating a new distance which encloses the original fade spheres.
+* Warn if propcombine or packing was disabled via command line.
 
 ## Bugfixes
 * Fix a compile failure if prop ropes were placed in a group with no connections.